Changes to index/test.html
Added a test.html to test removing the gutter from 3 images from Colt's example code. However he is using bootstrap 4.5 where 5 is the current version, so I had to fiddle with it in order to get it to work from his video, where no-gutters is the 4.5 version, g-0 is the version 5 tag unfortunately for me.