filter quiz
due to the difficulty I had with this quiz I decided to save the file from it to reference later. This one was particularly ridiculous for someone who was learning because the solution required nested functions, which we basically just touched on and didn't really expand on a lot in the lessons.